rename package and ditch boilerplate references
