+++ GalileoFront/app/build.gradle.kts
@@ -3,11 +3,11 @@ plugins {
 }
 
 android {
-    namespace = "com.example.galileomastermindboilerplate"
+    namespace = "com.gnsstracker.mainapp"
     compileSdk = 35
 
     defaultConfig {
-        applicationId = "com.example.galileomastermindboilerplate"
+        applicationId = "com.gnsstracker.mainapp"
         minSdk = 30
         targetSdk = 35
         versionCode = 1
