+++ GalileoFront/app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
         android:theme="@style/Theme.App.Starting"
         tools:targetApi="31">
         <activity
-            android:name=".ui.MainActivity"
+            android:name="com.gnsstracker.mainapp.ui.MainActivity"
             android:exported="true"
             android:screenOrientation="userPortrait"
             android:label="@string/app_name">
