+++ GalileoFront/app/src/main/java/com/gnsstracker/mainapp/MeasurementSerializer.java
@@ -1,4 +1,4 @@
-package com.example.galileomastermindboilerplate;
+package com.gnsstracker.mainapp;
 
 import android.location.GnssMeasurement;
