+++ GalileoFront/app/src/main/java/com/gnsstracker/mainapp/ServerHostname.java
@@ -1,4 +1,4 @@
-package com.example.galileomastermindboilerplate;
+package com.gnsstracker.mainapp;
 
 public class ServerHostname {
