+++ GalileoFront/app/src/main/res/layout/fragment_landing_page.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.LandingPage">
+    tools:context="com.gnsstracker.mainapp.ui.LandingPage">
 
     <!--ScrollView
         android:layout_width="match_parent"
