+++ GalileoFront/app/src/main/res/layout/fragment_satellite_filters_modal_pane_list_dialog.xml
@@ -4,7 +4,7 @@
     android:layout_height="wrap_content"
     xmlns:tools="http://schemas.android.com/tools"
     android:orientation="vertical"
-    tools:context=".ui.satellitelist.SatelliteListPage">
+    tools:context="com.gnsstracker.mainapp.ui.satellitelist.SatelliteListPage">
 
     <TextView
         android:id="@+id/textView1"
