+++ GalileoFront/app/src/main/res/values/themes.xml
@@ -1,12 +1,12 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
-    <style name="Theme.GalileoMastermindBoilerplate" parent="Theme.Material3.Light">
+    <style name="Theme.GNSSTracker" parent="Theme.Material3.Light">
         <item name="android:statusBarColor">#00123456</item>
         <item name="android:navigationBarColor">#00123456</item>
     </style>
     <style name="Theme.App.Starting" parent="Theme.SplashScreen">
         <item name="windowSplashScreenAnimatedIcon">@drawable/icon_splashscreen</item>
         <item name="windowSplashScreenAnimationDuration">2000</item>
-        <item name="postSplashScreenTheme">@style/Theme.GalileoMastermindBoilerplate</item>
+        <item name="postSplashScreenTheme">@style/Theme.GNSSTracker</item>
         <item name="windowSplashScreenBackground">#FFFFFF</item>
     </style>
 </resources>
